Fix storeOp for multisampled image
Before this fix, according to Best Practice Validation: "Attachment 0 in
the VkRenderPass is a multisampled image with 8 samples, but it uses
loadOp/storeOp which requires accessing data from memory. Multisampled
images should always be loadOp = CLEAR or DONT_CARE, storeOp =
DONT_CARE. This allows the implementation to use lazily allocated memory
effectively."

There doesn't seem to be a good reason for storeOp = STORE as we resolve
this attachment to a single-sample color attachment later.